refactoring: move const_to_pat code into its own submodule.
